Created get_all_tags endpoint, finished GetAllProfiles and ProfileForm frontend, changed type of phone_number in tables from INTEGER to BIGINT to allow user input larger number

+++ meal-ting-pot/migrations/003_user_profile_&_social_media_table.py
@@ -35,7 +35,7 @@ steps = [
         full_name VARCHAR(150) NOT NULL,
         email VARCHAR(40) NOT NULL,
         photo VARCHAR(1000) NOT NULL,
-        phone_number INTEGER NOT NULL,
+        phone_number BIGINT NOT NULL,
         address VARCHAR(300) NOT NULL,
         bio VARCHAR(1000) NOT NULL,
         availability BOOL NOT NULL,
